Refactor: Move ConfidentialMessageService.Result one level up
